+++ test/fuzz/Fuzz.t.sol
@@ -249,7 +249,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorUsd),
             address(mockERC20.stable1),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.stableToStableCollFactor,
             Constants.stableToStableLiqFactor
         );
@@ -257,7 +257,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorStable1),
             address(mockERC20.stable1),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.stableToStableCollFactor,
             Constants.stableToStableLiqFactor
         );
@@ -265,7 +265,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorToken1),
             address(mockERC20.stable1),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToStableCollFactor,
             Constants.tokenToStableLiqFactor
         );
@@ -274,7 +274,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorUsd),
             address(mockERC20.stable2),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.stableToStableCollFactor,
             Constants.stableToStableLiqFactor
         );
@@ -282,7 +282,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorStable1),
             address(mockERC20.stable2),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.stableToStableCollFactor,
             Constants.stableToStableLiqFactor
         );
@@ -290,7 +290,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorToken1),
             address(mockERC20.stable2),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToStableCollFactor,
             Constants.tokenToStableLiqFactor
         );
@@ -299,7 +299,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorUsd),
             address(mockERC20.token1),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToStableCollFactor,
             Constants.tokenToStableLiqFactor
         );
@@ -307,7 +307,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorStable1),
             address(mockERC20.token1),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToStableCollFactor,
             Constants.tokenToStableLiqFactor
         );
@@ -315,7 +315,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorToken1),
             address(mockERC20.token1),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToTokenCollFactor,
             Constants.tokenToTokenLiqFactor
         );
@@ -324,7 +324,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorUsd),
             address(mockERC20.token2),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToStableCollFactor,
             Constants.tokenToStableLiqFactor
         );
@@ -332,7 +332,7 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorStable1),
             address(mockERC20.token2),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToStableCollFactor,
             Constants.tokenToStableLiqFactor
         );
@@ -340,29 +340,29 @@ abstract contract Fuzz_Test is Base_Test {
             address(creditorToken1),
             address(mockERC20.token2),
             0,
-            type(uint128).max,
+            type(uint112).max,
             Constants.tokenToTokenCollFactor,
             Constants.tokenToTokenLiqFactor
         );
 
         registryExtension.setRiskParametersOfPrimaryAsset(
-            address(creditorUsd), address(mockERC721.nft1), 0, type(uint128).max, 0, 0
+            address(creditorUsd), address(mockERC721.nft1), 0, type(uint112).max, 0, 0
         );
         registryExtension.setRiskParametersOfPrimaryAsset(
-            address(creditorStable1), address(mockERC721.nft1), 0, type(uint128).max, 0, 0
+            address(creditorStable1), address(mockERC721.nft1), 0, type(uint112).max, 0, 0
         );
         registryExtension.setRiskParametersOfPrimaryAsset(
-            address(creditorToken1), address(mockERC721.nft1), 0, type(uint128).max, 0, 0
+            address(creditorToken1), address(mockERC721.nft1), 0, type(uint112).max, 0, 0
         );
 
         registryExtension.setRiskParametersOfPrimaryAsset(
-            address(creditorUsd), address(mockERC1155.sft1), 1, type(uint128).max, 0, 0
+            address(creditorUsd), address(mockERC1155.sft1), 1, type(uint112).max, 0, 0
         );
         registryExtension.setRiskParametersOfPrimaryAsset(
-            address(creditorStable1), address(mockERC1155.sft1), 1, type(uint128).max, 0, 0
+            address(creditorStable1), address(mockERC1155.sft1), 1, type(uint112).max, 0, 0
         );
         registryExtension.setRiskParametersOfPrimaryAsset(
-            address(creditorToken1), address(mockERC1155.sft1), 1, type(uint128).max, 0, 0
+            address(creditorToken1), address(mockERC1155.sft1), 1, type(uint112).max, 0, 0
         );
 
         vm.stopPrank();
